backend/pyproject.toml
[project]
name = "welfare-support-backend"
version = "0.1.0"
description = "Welfare Support Management System Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "anthropic>=0.34.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "aiofiles>=23.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []
